finish rename mdk to fdk
