+++ tests/metagen/typegraphs/fdk_py_templates/struct.py.jinja
