+++ tests/metagen/typegraphs/identities/rs/build.sh
@@ -9,7 +9,7 @@ TARGET=wasm32-wasi
 cargo build --target $TARGET # --release
 # wasm-opt --enable-bulk-memory -Oz ./target/$TARGET/release/rust.wasm -o ./target/rust-component.wasm.opt
 wasm-tools component new \
-  ./target/$TARGET/debug/metagen_identities_mdk.wasm \
+  ./target/$TARGET/debug/metagen_identities_fdk.wasm \
   -o ./target/rust-component.wasm \
   --adapt wasi_snapshot_preview1=$ADAPTOR
 # wasm-tools component wit target/rust-component.wasm
