+++ tests/metagen/typegraphs/identities/rs/fdk.rs
