+++ tests/metagen/typegraphs/identities/rs/lib.rs
@@ -1,8 +1,8 @@
 // Copyright Metatype OÜ, licensed under the Elastic License 2.0.
 // SPDX-License-Identifier: Elastic-2.0
 
-mod mdk;
-pub use mdk::*;
+mod fdk;
+pub use fdk::*;
 
 init_mat! {
     hook: || {
