+++ tests/metagen/typegraphs/identities/ts/fdk.ts
