+++ tests/metagen/typegraphs/sample/rs/Cargo.toml
@@ -1,15 +1,15 @@
-package.name = "sample_mdk"
+package.name = "sample_fdk"
 package.edition = "2021"
 package.version = "0.0.1"
 
 [dependencies]
 serde = { version = "1.0.203", features = ["derive"] }
 serde_json = "1.0.117"
-reqwest = { version = "0.12", features = ["blocking","json"] }
+reqwest = { version = "0.12", features = ["blocking", "json"] }
 tokio = { version = "1", features = ["rt-multi-thread"] }
 
 # The options after here are configured for crates intended to be
 # wasm artifacts. Remove them if your usage is different
 [[bin]]
-name = "sample_mdk"
+name = "sample_fdk"
 path = "main.rs"
