+++ tests/metagen/typegraphs/sample/ts/client.ts
@@ -656,11 +656,6 @@ const nodeMetas = {
       ],
     };
   },
-  Func27(): NodeMeta {
-    return {
-      ...nodeMetas.Post(),
-    };
-  },
   User(): NodeMeta {
     return {
       subNodes: [
@@ -675,14 +670,14 @@ const nodeMetas = {
       ...nodeMetas.User(),
     };
   },
+  Func27(): NodeMeta {
+    return {
+      ...nodeMetas.Post(),
+    };
+  },
   Func25(): NodeMeta {
     return {
       ...nodeMetas.scalar(),
-    };
-  },
-  Func24(): NodeMeta {
-    return {
-      ...nodeMetas.Post(),
     };
   },
   Func28(): NodeMeta {
@@ -703,15 +698,20 @@ const nodeMetas = {
       },
     };
   },
-};
-export type Object21 = {
-  id: string;
+  Func24(): NodeMeta {
+    return {
+      ...nodeMetas.Post(),
+    };
+  },
 };
 export type StringUuid4 = string;
 export type Post = {
   id: StringUuid4;
   slug: string;
   title: string;
+};
+export type Object21 = {
+  id: string;
 };
 export type StringEmail5 = string;
 export type Post7 = Array<Post>;
