+++ tests/runtimes/wasm_wire/rust/fdk.rs
