+++ tools/Dockerfile.dockerignore
@@ -7,7 +7,7 @@
 !src/common
 !src/meta-cli/Cargo.toml
 !src/metagen/Cargo.toml
-!src/metagen/src/mdk_rust/static/Cargo.toml
+!src/metagen/src/fdk_rust/static/Cargo.toml
 !src/mt_deno
 !src/pyrt_wit_wire
 !src/typegate/
